+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestBody.java
@@ -1,0 +1,4 @@
+package org.apache.coyote.http11.request;
+
+public class RequestBody {
+}
